requirements.txt
ipykernel
openai
langchain
langchain-community
langchain-openai
pypdf
flask
python-dotenv
requests
PyPDF2
chromadb
lxml
langchain-text-splitters
sentence-transformers
langchain-chroma